requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Multimodal Processing
opencv-python>=4.8.0
Pillow>=10.0.0
librosa>=0.10.0
scipy>=1.10.0
soundfile>=0.12.0
moviepy>=1.0.3

# Ethics and AI Safety
datasets>=2.14.0
evaluate>=0.4.0
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0

# Constitutional AI and RLHF
peft>=0.4.0
trl>=0.7.0
wandb>=0.15.0
tensorboard>=2.13.0

# API and Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
httpx>=0.24.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0

# Database and Storage
redis>=4.5.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# Testing and Development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0
sentry-sdk>=1.28.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# Meta Integration (Placeholder for internal packages)
# meta-llama-client>=1.0.0  # Internal Meta package
# meta-multimodal-api>=1.0.0  # Internal Meta package
# meta-ethics-toolkit>=1.0.0  # Internal Meta package